// frontend/src/app/models/sportistaObject.ts
export class SportistaObject {
    ime: string ; // String ime = new String("Aleksandar")
    prezime: string;
    nacionalnost: string;
    finalista: boolean = false;
    zlato: boolean = false;
    srebro: boolean = false;
    bronza: boolean = false;
    novoKolo: boolean = false;
    rez1: string = " ";
    rez2: string = " ";
    rez3: string = " ";
    rez4: string = " ";
    rez5: string = " ";
    rez6: string = " ";
    rezNoviKrug: string = " ";
} 